remove unused imports and sperate rule interface in a sperated file
