+++ src/app/core/models/skolera-interfaces.model.ts
@@ -1,98 +1,47 @@
-import { Rule } from "./rules-interfaces.model";
 
 export interface PaginationData {
-    current_page: number ;
-    next_page: number;
-    prev_page: number;
-    total_count: number;
-    total_pages: number
-}
-export interface Employee {
-  id?: number
-    number: string;
-    name: string;
-    department_name: number;
-    biometric_id: number;
-    isSelected? : boolean
-    time_group_schedule?: TimeGroupSchedule[]
-  
-  }
-
-export class TimeGroup {
-    
-    constructor(){ }
-    name: string;
-    id?: number;
-    group_type: string;
-    number_of_employees?: number;
-    employees?: Employee [];
-    employees_attributes?: EmployeesAttributes;
-    time_group_schedule_attributes?: {
-      schedule_days_attributes? :  TimeGroupSchedule[],
-      schedule_days? :  TimeGroupSchedule[]
-    } ;
-     time_group_schedule?: {
-      schedule_days? :  TimeGroupSchedule[]
-      schedule_days_attributes? :  TimeGroupSchedule[],
-    } ;
-    rule?: Rule
-
-} 
-export interface EmployeesAttributes{
-  id: number,
-	time_group_schedule_attributes: {
-			schedule_days_attributes :  TimeGroupSchedule[]
-    }
+  current_page: number;
+  next_page: number;
+  prev_page: number;
+  total_count: number;
+  total_pages: number
 }
 
 
-export class TimeGroupSchedule {
-    day: string;
-    clock_in?: string | null;
-    clock_out?: string | null;
-    is_off?: boolean;
-    calculate_bridging?: boolean;
-    invalidTime?: boolean;
-}
-  export interface CurrentUser {
-    id: number;
-    actable_id: number;
-    actable_type: string;
-    actions: [];
-    avatar_url: string
-    child_id?: number;
-    children?: null;
-    city?: string;
-    country?: string;
-    dateofbirth?: Date;
-    email?: string
-    firstname: string;
-    gender: string;
-    hide_grades?: boolean;
-    home_address?: string;
-    is_active?: boolean;
-    last_sign_in_at?: Date;
-    lastname: string;
-    locale: string;
-    middlename?: string;
-    name: string;
-    role: {};
-    parent_id?: number;
-    password: string;
-    phone?: string;
-    realtime_ip?: string;
-    school_name?: string;
-    secondary_address?: string;
-    secondary_phone?: string;
-    show_school_fees?: string;
-    thumb_url?: string;
-    unseen_notifications?: number;
-    user_type?: string;
-    username: string;
+export interface CurrentUser {
+  id: number;
+  actable_id: number;
+  actable_type: string;
+  actions: [];
+  avatar_url: string
+  child_id?: number;
+  children?: null;
+  city?: string;
+  country?: string;
+  dateofbirth?: Date;
+  email?: string
+  firstname: string;
+  gender: string;
+  hide_grades?: boolean;
+  home_address?: string;
+  is_active?: boolean;
+  last_sign_in_at?: Date;
+  lastname: string;
+  locale: string;
+  middlename?: string;
+  name: string;
+  role: {};
+  parent_id?: number;
+  password: string;
+  phone?: string;
+  realtime_ip?: string;
+  school_name?: string;
+  secondary_address?: string;
+  secondary_phone?: string;
+  show_school_fees?: string;
+  thumb_url?: string;
+  unseen_notifications?: number;
+  user_type?: string;
+  username: string;
 }
 
-export interface Department{
-  name: string,
-  id: number,
-  code: string
-}
