+++ src/app/main/time-groups/edit-time-group/edit-time-group.component.ts
@@ -1,11 +1,11 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { Department, Employee, PaginationData, TimeGroup, TimeGroupSchedule } from '@core/models/skolera-interfaces.model';
+import { MatDialog } from '@angular/material/dialog';
+import { TimeGroup, TimeGroupSchedule } from '@core/models/time-groups-interface.model';
+import { Department, Employee } from '@core/models/employees-interface.model';
+import { PaginationData } from '@core/models/skolera-interfaces.model';
 import { TimeGroupsSerivce } from '@skolera/services/time-groups.services';
-import { HeaderComponent } from '../../header/header.component';
 import { AppNotificationService } from '@skolera/services/app-notification.service';
 import { TranslateService } from '@ngx-translate/core';
-import * as moment from 'moment';
 import { NgForm } from '@angular/forms';
 import { EmployeesSerivce } from '@skolera/services/employees.services';
 import { Subscription } from 'rxjs';
